support_backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
flask-marshmallow>=1.2.0
marshmallow>=3.21.0
marshmallow-sqlalchemy>=1.0.0
flasgger>=0.9.7.1
python-dotenv>=1.0.0
Werkzeug>=3.0.0
bcrypt>=4.1.0
bleach>=6.1.0
Flask-Limiter>=3.5.0
email-validator>=2.1.0

pytest>=8.0.0
pytest-cov>=5.0.0
